+++ client/src/components/index.ts
@@ -1,0 +1,4 @@
+import PostPreview from './PostPreview'
+import SearchBar from './SearchBar'
+
+export { PostPreview, SearchBar }
