+++ client/src/pages/index.ts
@@ -1,0 +1,8 @@
+import HomePage from './HomePage'
+import Login from './Login'
+import NewPostPage from './NewPostPage'
+import NotFound from './NotFound'
+import PostPage from './PostPage'
+import Profile from './Profile'
+
+export {HomePage, Login, NewPostPage, NotFound, PostPage, Profile}
